cleaned up some internal enum and toString operators.
