+++ src/core/scenes/SettingsScene.cpp
@@ -123,7 +123,7 @@ void SettingsScene::Update(float dt)
 
     if (m_hasPendingTransition)
     {
-        CT_LOG_INFO("SettingsScene Requesting Scene Change to '{}'", ToString(m_requestedScene));
+        CT_LOG_INFO("SettingsScene Requesting Scene Change to '{}'", SceneIDToString(m_requestedScene));
         m_hasPendingTransition = false;
         SceneTransitionManager::Instance().ForceFullyOpaque();
         SceneManager::Instance().RequestSceneChange(m_requestedScene);
